(refactor: controller-service 정리)

+++ src/main/java/com/walkbuddies/backend/dto/parkservice/ParkDto.java
@@ -17,7 +17,7 @@ public class ParkDto {
     private String sportFacility;
     private String convenienceFacility;
 
-    private float distance;
+    private Double distance;
 
     public static ParkDto convertToDto(ParkEntity parkEntity) {
         return ParkDto.builder()
